+++ packages/scheduler/src/SchedulerFeatureFlags.js
@@ -10,3 +10,4 @@ export const enableSchedulerDebugging = false;
 export const enableIsInputPending = false;
 export const requestIdleCallbackBeforeFirstFrame = false;
 export const requestTimerEventBeforeFirstFrame = false;
+export const enableMessageLoopImplementation = false;
